+++ mkbak/__init__.py
@@ -1,1 +1,1 @@
-version: str = "1.0.3"
+version: str = "1.1.0"